+++ CHANGELOG.md
@@ -4,12 +4,17 @@ All notable changes to Proxmox VE for WHMCS will be documented in this file.
 ## [1.2.2b] - Not yet released
 
 ### 🚀 Feature
-- Power Actions now available in Admin Area as well as Client Area
+- Power Actions: Now available in Admin Area as well as Client Area
 - (Note: Suspend/Unsuspend/Terminate remain admin-only functions)
 
 ### 💅 Polish
 - Client Area: Power Action wording amended (Soft Stop, Hard Stop)
 - Admin, Module Config: Explain what the VNC Secret field is about
+- Admin, Module Config: House-keeping to design, Support/Health tab
+
+### 🐛 Bug Fix
+- Admin, Create Service: Fails if Plan/Pool not assigned in WHMCS (#36)
+- Client, VNC: Fails early if VNC Secret is not set or adequate (#27)
 
 ## [1.2.1b] - 2023-06-19